+++ CHANGE_LOG.md
@@ -9,3 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Initial release
+
+### Changed
+
+- The `Lumberjack::Logger#tagged` method stores tags on the "tags" attribute instead of the "tagged" attribute as it did in older versions of the `lumberjack` gem.